requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
bcrypt==4.0.1
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
docxcompose==1.4.0
docxtpl==0.16.7
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
oauth2client==4.1.3
openai==0.27.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyDrive==1.3.1
PyJWT==2.6.0
pymongo==4.3.3
pyparsing==3.0.9
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.2.4
stripe==5.4.0
tqdm==4.65.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
yarl==1.8.2
